+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [0.1.0] - 2019-05-04
+### Added
+- one basic example
+- support to validate on change and on mount
+- isValid method
+- ability to touch and untouch fields
+- ability to get form values
+- ability to force validate
+- README
+
 ## [0.0.1-beta.1] - 2019-04-09
 ### Added
 - initial version